Implement pay service , add epired  field to OrderPaymentProces,  renaming

+++ DataFlow/Models/OrderPaymentDetail.cs
@@ -11,7 +11,7 @@ namespace DataTier.Models
         public int Id { get; set; }
         public int OrderId { get; set; }
         public Order Order { get; set; }
-        public int Amaunt { get; set; }
+        public decimal Amaunt { get; set; }
         public string ExternalID { get; set; }
         public string CartType { get; set; }
         public string CartNubmer { get; set; }
